522: v1.8: add route api and fix bug redirect(url_for())
